+++ .gitignore
@@ -7,7 +7,7 @@
 *.pyc
 __pycache__/
 local_settings.py
-db.sqlite3
+server/db.sqlite3
 db.sqlite3-journal
 media
