+++ server/apps/users/views.py
@@ -1,0 +1,3 @@
+from django.shortcuts import render
+
+# Create your views here.
